in this project i created a VPC
I connected an internet gateway to the VPC which allowed the resources in the VPC to communicate to the internet. The route table for my public subnet has a route to the internet gateway. The NAT Gateway is put inside of the public subnet so it can reach the internet gateway which in turn provides the private subnet with internet access. The security groups allow traffic to and from the internet